+++ Viento.java
@@ -1,7 +1,7 @@
 public class Viento extends Producto {
     int largo; 
     
-    public Viento(String c, double precio, String mar, String mo, String n, String mat, int p, int largo){
+    public Viento(String c, double precio, String mar, String mo, String n, String mat, double p, int largo){
         super(c, precio, mar, mo, n, mat, p);
         this.largo =  largo; 
     }
